refactor(onboarding): move command orchestration into tau-onboarding
